chore(release): record batch one fixes, changelog and consolidated validation
Finalize regression coverage and update download API documentation. Batch one is complete; later maintenance groups wait for the user to continue.
