+++ frontend/src/utils/translationParagraphs.ts
@@ -1,8 +1,27 @@
 // Normalize RSS fragments into paragraphs without separating inline emphasis or links.
 export function wrapOrphanedTextNodes(container: Element): void {
   const inlineTags = new Set([
-    'A', 'SPAN', 'STRONG', 'EM', 'B', 'I', 'U', 'S', 'SMALL', 'SUB', 'SUP',
-    'BR', 'CODE', 'KBD', 'MARK', 'ABBR', 'TIME', 'DEL', 'INS', 'CITE', 'Q',
+    'A',
+    'SPAN',
+    'STRONG',
+    'EM',
+    'B',
+    'I',
+    'U',
+    'S',
+    'SMALL',
+    'SUB',
+    'SUP',
+    'BR',
+    'CODE',
+    'KBD',
+    'MARK',
+    'ABBR',
+    'TIME',
+    'DEL',
+    'INS',
+    'CITE',
+    'Q',
   ]);
   const blocks = [container, ...Array.from(container.querySelectorAll('div,section,article'))];
   for (const block of blocks) {
@@ -17,7 +36,10 @@ export function wrapOrphanedTextNodes(container: Element): void {
       run = [];
     };
     for (const node of Array.from(block.childNodes)) {
-      if (node.nodeType === Node.TEXT_NODE || (node instanceof Element && inlineTags.has(node.tagName))) {
+      if (
+        node.nodeType === Node.TEXT_NODE ||
+        (node instanceof Element && inlineTags.has(node.tagName))
+      ) {
         run.push(node);
       } else {
         flush();
